+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# 0.11.0
+
+## Added
+
+-   Add tectonic backend to compile vesti. In this version, it can be installed
+    with enabling `tectonic-backed` feature. Here is the command line argument to
+    install this feature
+    ```console
+    $ cargo install vesti --features=tectonic-backed
+    ```
+-   The new keyword `importfile*` copies the file into `vesti-cache` folder. For instance, if one write
+    ```vesti
+    importfile* (./foo/bar.txt)
+    ```
+    then vesti copies `bar.txt` inside in the `./foo` directory into `vesti-cache`.
+    Especially, one can write
+    ```vesti
+    importfile* (@/foo/bar.txt)
+    ```
+    The special directory `@` refers to the OS specific config directory. For more information about config directory location, see [here](https://docs.rs/dirs/5.0.1/dirs/fn.config_dir.html).
+
 # 0.10.1
 
 ## Changed